requirements.txt
flask==3.0.0
werkzeug==3.0.1
twilio==8.10.0
openai==1.3.5
sqlalchemy==2.0.23
python-dotenv==1.0.0
gunicorn==21.2.0
